% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_mat_func.R
\name{condition_mat}
\alias{condition_mat}
\title{Constraint matrix in quadratic optimization problem}
\usage{
condition_mat(event_vec, num_event_time)
}
\arguments{
\item{event_vec}{vector containing information if a subject is at risk or if an event happens. If n are the number of subjects and m the number of event times, then event_vec has length n*m}

\item{num_event_time}{number of event}
}
\value{
matrix
}
\description{
calculates the matrix which defines the constraints in the SVHM algorithm
}
