% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_svhm.R
\name{create_svhm}
\alias{create_svhm}
\title{Train SVHM}
\usage{
create_svhm(
  df,
  covariates,
  cross_validation_val,
  cost_grid,
  varName_cencored,
  varName_futime,
  k = 3,
  test_size = 0.2,
  opt = "osqp",
  gamma_squared = 0.5
)
}
\arguments{
\item{df}{data frame}

\item{covariates}{vector of name of covariates}

\item{cross_validation_val}{number of subset to use for cost optimization}

\item{cost_grid}{grid of all cost parameter to be optoimzed uponl}

\item{varName_cencored}{name of variable in df that indicates cencoring}

\item{varName_futime}{name of variable in df that indicates event time}

\item{k}{integer of how many nearest event times are used to predict the event time (default is 3)}

\item{test_size}{size of final test set in precent}

\item{opt}{which quadratic optimization is used (\code{opt='mosek'} or \code{opt='osqp'})}

\item{gamma_squared}{width of gaussian kernel}
}
\value{
{trained model with
         \code{$e_vec} vector indicating vector containing information if a subject is at risk or if an event happens. If n are the number of subjects and m the number of event times, then event_vec has length n*m,
         \code{$k_mat} kernel matrix,
         \code{$sol} calculated optimal solution,
         \code{$t_predict} test dataset with risk scores \code{risk} and \code{t_predict},
         \code{$p_corr} pearson correlation of the predicted times
}
}
\description{
predicts the event times of a given dataset using cross validation for the cost parameter. Final model includes predicted event times as well as parameters to predict new event times from subject who are not in df.
All values of covariates are first normalized to the intervall [0,1] before the SVHM algorithm is applied. The cost parameter for the final model is chosen with the best pearson correlation.
}
\note{
The mosek package requires a license
}
\examples{
{

library(KMsurv)
library(SVHM)

##############
# Parameters #
##############

gamma_squared <- 100
k <- 1
cross_validation_val <- 3
test_size=.3
cost_grid <- 2^c(-6:6)

covariates <- c('z7')

######################
#  Model prediction  #
######################

data(bmt)

model <- create_svhm(bmt, covariates, cross_validation_val, cost_grid, varName_cencored="d3", varName_futime = "t2", k=k, test_size=test_size, opt='osqp', gamma_squared=gamma_squared)
}

}
