% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_time_svhm.R
\name{create_time_svhm}
\alias{create_time_svhm}
\title{Train Time Dependent SVHM}
\usage{
create_time_svhm(
  df,
  covariates,
  cost,
  varName_cencored,
  varName_futime,
  start_interval,
  end_interval,
  test_size = 0.3,
  opt = "osqp",
  gamma_squared = 0.5
)
}
\arguments{
\item{df}{data frame}

\item{covariates}{vector of name of covariates}

\item{cost}{cost parameter to be used}

\item{varName_cencored}{name of variable in df that indicates cencoring}

\item{varName_futime}{name of variable in df that indicates event time}

\item{test_size}{size of final test set in precent}

\item{opt}{which quadratic optimization is used (\code{opt='mosek'} or \code{opt='osqp'})}

\item{gamma_squared}{width of gaussian kernel}
}
\value{
{trained model with
         \code{$e_vec} vector indicating if an event happens at each event time
         \code{$sol} calculated optimal solution for each event time
         \code{$train} train dataset with risk scores
         \code{$test} test dataset with risk scores
         \code{cost} cost parameter
}
}
\description{
Calculates the Risk score and the value of the prediction function for each individual in the data set.
}
\note{
In contrast to the \code{create_svhm()} function this function does not predict event times!
}
\examples{
{

library(timereg)
library(SVHM)

##############
# Parameters #
##############

opt <- "osqp"
gamma_squared <- 200
test_size=.3
cost <- 16

######################
#  Model prediction  #
######################

data(csl)

time_model <- create_time_svhm(csl, c("sex", "age"), cost, varName_cencored='dc', varName_futime='eventT', start_interval='lt, end_interval='rt, test_size=test_size, opt=opt, gamma_squared=gamma_squared)
}

}
