% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDataPartition.R
\name{createListPartition}
\alias{createListPartition}
\title{createListPartition}
\usage{
createListPartition(l, cross_validation_val, test_size = 0.2)
}
\arguments{
\item{l}{list of data}

\item{cross_validation_val}{number of sets for k-fold cross validation}

\item{test_size}{size of test set (default=.8)}
}
\value{
{partitioned list
}
}
\description{
partitions a List into a test set and cross validation sets. createDataPartition() is not randomized, therefore the list should be randomized before creating the partition!
}
\examples{
{
# Example with the preloaded mtcars dataset
l<-list("A", "B", "C", "D", "E", "F", "G", "H", "I", "J")
partition <- SVHM:::createListPartition(l, 3, .1)
}

}
