% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_at_time.R
\name{data_at_time}
\alias{data_at_time}
\title{data_at_time}
\usage{
data_at_time(i, j, df, times)
}
\arguments{
\item{i}{index of individuals}

\item{j}{index of event time}

\item{df}{list of dataframes of the individuals}

\item{times}{dataframe of event times}
}
\value{
row of data for i-th individual if the individual is still under risk, otherwise return row of NA.
}
\description{
Retrieves the relevant data for each individual at every event time.
}
