% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize}
\usage{
normalize(df, covariates)
}
\arguments{
\item{df}{dataframe}

\item{col}{columns to be normalized}
}
\value{
normalized columns of dataframe
}
\description{
normalizes a vector
}
\examples{
{
Example with the preloaded mtcars dataset
SVHM:::normalize(mtcars, c('disp',  'hp'))
}

}
