% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sol_mosek_func.R
\name{opt_sol_mosek}
\alias{opt_sol_mosek}
\title{Optimal solution of SVHM}
\usage{
opt_sol_mosek(optimizazion_data, num_event_times, cost)
}
\arguments{
\item{num_event_times}{number of event times in the training dataset}

\item{cost}{cost parameter of the support vector machine of type numeric}

\item{optimization_data}{all values needed for optimization in a list with order (risk_vector, adapted_kernel_matrix, cond_mat, weight_vec)}
}
\value{
optimal solution for the SVHM
}
\description{
Uses the Rmosek package to solve the quadratic optimization problem defined by SVHM.
}
