% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sol_osqp_func.R
\name{opt_time_sol_osqp}
\alias{opt_time_sol_osqp}
\title{Optimal solution of SVHM}
\usage{
opt_time_sol_osqp(e_vec, k_mat, w_vec, cost)
}
\arguments{
\item{cost}{cost parameter of the support vector machine of type numeric}

\item{optimization_data}{all values needed for optimization in a list with order (risk_vector, adapted_kernel_matrix, cond_mat, weight_vec)}

\item{num_event_times}{number of event times in the training dataset}
}
\value{
optimal solution for the SVHM
}
\description{
Uses the osqp package to solve the quadratic optimization problem defined by SVHM.
}
