% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_data_func.R
\name{optimization_data}
\alias{optimization_data}
\title{Optimization Data}
\usage{
optimization_data(
  covariates,
  training_dataset,
  ordered_event_times,
  gamma_squared
)
}
\arguments{
\item{covariates}{dataset of covariates of the subjects in a dataset}

\item{training_dataset}{data frame representing the trainings dataset}

\item{ordered_event_times}{data frame of all event times ordered in ascending order}

\item{gamma_squared}{width of the kernel}
}
\value{
{
\item risk_vector            vector representing at which event times the subjects are under risk
\item adapted_kernel_matrix  matrix on which quadratic optimization will be performed
\item cond_mat               matrix representing the constraints of the optimization problem
\item weight_vec             vector of weights at any event time for all subjects
\item kernel_matrix          Gram matrix of covariates
}
}
\description{
calculates all needed values to execute quadratic optimization in SVHM
}
