% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_data_func.R
\name{optimization_time_data}
\alias{optimization_time_data}
\title{Time Dependent Optimization Data}
\usage{
optimization_time_data(covariates, mat_train, event_time, gamma_squared = 0.5)
}
\arguments{
\item{covariates}{dataset of covariates of the subjects in a dataset}

\item{mat_train}{matrix of all individuals under risk at the event time}

\item{event_time}{event time for which data is calculated}

\item{gamma_squared}{width of gaussian kernel}
}
\value{
{List
\code{$adap_k_mat}  matrix on which quadratic optimization will be performed
\code{$w_vec}       vector of weights at any event time for all subjects
\code{$k_mat}       Gram matrix of covariates
\code{$e_vec}       vector indicating vector containing information if a subject experiences an event.
}
}
\description{
calculates all needed values to execute quadratic optimization for the time dependent SVHM at the given event time.
}
