% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_func.r
\name{radial_kernel}
\alias{radial_kernel}
\title{Gaussian Kernel}
\usage{
radial_kernel(x, y, gamma_squared)
}
\arguments{
\item{x}{first input vector}

\item{y}{second input vector}

\item{gamma_squared}{width of the kernel}
}
\value{
gaussian kernel value
}
\description{
calculates the Gaussian Kernel value of two inputs
}
\examples{
{
x <- runif(n=10)
y <- runif(n=10)
radial_kernel(x,y,.5)
}

}
