% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_func.r
\name{radial_kernel_mat}
\alias{radial_kernel_mat}
\title{Gaussian Kernel Matrix}
\usage{
radial_kernel_mat(covariates, gamma_squared)
}
\arguments{
\item{covariates}{dataset of covariates of the subjects in a dataset}

\item{gamma_squared}{width of the kernel}
}
\value{
gaussian kernel matrix
}
\description{
calculates the gaussian kernel value of the covariates with each other. calculated matrix will be symmetric
}
\examples{
{
# Example with the preloaded mtcars dataset
covariates <- subset( mtcars, select = c('drat', 'wt') )
SVHM:::radial_kernel_mat(covariates,.5)
}

}
