% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_score_func.r
\name{risk_score_training}
\alias{risk_score_training}
\title{Training risk scores}
\usage{
risk_score_training(
  gamma_sol,
  kernel_mat,
  event_vec,
  num_event_times,
  training_set_size
)
}
\arguments{
\item{gamma_sol}{optimal solution of the SVHM}

\item{kernel_mat}{Gram matrix of the covariates}

\item{num_event_times}{number of event times that occour in the training data set}
}
\value{
vector of risk scores for all training subjects
}
\description{
calculates the risk scores for all individuals in the training dataset.
}
\note{
The calculated risk scores are not the actual risk scores defined by the Risk function but the induce an ordering of the risk scores. For detailed information see reference
}
\references{
Wang, Y., Chen, T., and Zeng, D. Support vector hazards machine: A counting process framework for learning risk scores for censored outcomes. Journal of Machine Learning Research, 17(167):1-37, 2016
}
