% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_score_func.r
\name{risk_time_score}
\alias{risk_time_score}
\title{risk time scores}
\usage{
risk_time_score(
  gamma_sol,
  event_vec,
  weight_vec,
  v,
  covariates_train,
  n,
  gamma_squared = 0.5
)
}
\arguments{
\item{gamma_sol}{optimal solution of the SVHM}

\item{event_vec}{vector containing information of the training if a subject experiences an event happens.}

\item{weight_vec}{vector containing weigths}

\item{v}{covariates of the individual for which the risk is to be calculated}

\item{covariates_train}{dataset of covariates of the subjects in the training dataset}

\item{n}{number of individuals in the training dataset}

\item{gamma_squared}{width of gaussian kernel}
}
\value{
{\code{$f_at_j} decision function at j
         \code{$risk_at_j} risk score of the individualat time j
         }
}
\description{
calculates the risk scores for one individual with the help of the calculated optimal solution to the quadratic programming problem of time dependent SVHM at time j.
}
