% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_score_func.r
\name{risk_time_score_training}
\alias{risk_time_score_training}
\title{Training risk time scores}
\usage{
risk_time_score_training(kernel_mat, event_vec, weight_vec, f_vec, n)
}
\arguments{
\item{kernel_mat}{Gram matrix of the covariates}

\item{event_vec}{vector containing information of the training if a subject experiences an event happens.}

\item{weight_vec}{vector containing weigths}

\item{f_vec}{optimal decision function}

\item{n}{number of individuals in the training dataset}
}
\value{
vector of risk scores for all training subjects
}
\description{
calculates the risk scores for all individuals in the training dataset for the time dependent SVHM.
}
