% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_svhm.R
\name{train_svhm}
\alias{train_svhm}
\title{Train SVHM}
\usage{
train_svhm(train, test, covariates, cost, gamma_squared, k = 3, opt = "osqp")
}
\arguments{
\item{train}{training dataset}

\item{test}{test dataset}

\item{cost}{cost parameter of the support vector machine of type numeric}

\item{gamma_squared}{width of the kernel}

\item{k}{integer of how many nearest event times are used to predict the event time (default is 3)}

\item{opt}{which quadratic optimization is used (\code{opt='mosek'} or \code{opt='osqp'})}
}
\value{
{trained model with
         \code{$e_vec} vector indicating vector containing information if a subject is at risk or if an event happens. If n are the number of subjects and m the number of event times, then event_vec has length n*m,
         \code{$k_mat} kernel matrix,
         \code{$sol} calculated optimal solution,
         \code{$t_predict} test dataset with risk scores \code{risk} and \code{t_predict},
         \code{$p_corr} pearson correlation of the predicted times
}
}
\description{
Uses the Rmosek or osqp package to train the SVHM on a given training and test set.
}
