% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_mat_func.r
\name{weight_mat}
\alias{weight_mat}
\title{Weight matrix}
\usage{
weight_mat(training_dataset, ordered_event_times)
}
\arguments{
\item{training_dataset}{data frame representing the training data}

\item{ordered_event_times}{data frame of all event times ordered in ascending order}
}
\value{
matrix storing all weights for every individual
}
\description{
calculates the weights for every individual in the training dataset at every event time. If an individual experiences an event the weight is given by the ratio of at risk subjects with no event to all at risk subjects.
If no event is experienced the weight is given by the ratio of one over all at risk subjects.
}
